+++ tests/multi_configuration_project_configuration/conftest.py
@@ -2,7 +2,7 @@ import pytest
 import allure
 
 from pages.multi_config_project_config_page import MultiConfigProjectConfigPage
-from tests.multi_config_project_configuration.data import ProjectToggle, project_name
+from tests.multi_configuration_project_configuration.data import ProjectToggle, project_name
 
 
 @pytest.fixture
@@ -12,7 +12,7 @@ def multi_config_project_config_page(new_item_page):
 
 @pytest.fixture
 def multi_config_project_with_description(new_item_page):
-    from tests.multi_config_project_configuration.data import project_name, description_text
+    from tests.multi_configuration_project_configuration.data import project_name, description_text
     config_page = new_item_page.create_new_multi_config_project(project_name)
     return config_page.set_description(description_text, project_name)
